requirements.txt
streamlit
torch
torchvision
efficientnet-pytorch
opencv-python
matplotlib
numpy
Pillow
